feat: Add initial NixOS configuration
Adds NixOS configuration to configure my base system with all sorts of things,
including - but not limited to - i3, emacs, FF57, connman and so on.

Not all of these things are fully configured yet (especially in the desktop
environment) but I'm hoping to get more of that done in the future.